requirements.txt
pylint==2.17.5
pyyaml==6.0.2